--- src/components/svg/arrowDown.tsx
@@ -1,6 +1,0 @@
-export const ArrowDown = () => (
-    <svg width="9" height="9" viewBox="0 0 9 9" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path d="M1 0.999959L7.50004 7.5" stroke="#B1454F" stroke-width="1.4" stroke-linecap="round" stroke-linejoin="round" />
-        <path d="M3.13037 7.5H7.50015V3.13022" stroke="#B1454F" stroke-width="1.4" stroke-linecap="round" stroke-linejoin="round" />
-    </svg>
-);
