--- src/components/svg/arrowUp.tsx
@@ -1,6 +1,0 @@
-export const ArrowUp = () => (
-    <svg width="9" height="9" viewBox="0 0 9 9" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path d="M1.38184 8.22227L8.14066 1" stroke="#22C55E" stroke-width="1.4" stroke-linecap="round" stroke-linejoin="round" />
-        <path d="M3.59766 1H8.1414V5.85531" stroke="#22C55E" stroke-width="1.4" stroke-linecap="round" stroke-linejoin="round" />
-    </svg>
-);
