+++ src/components/svg/audioWave.tsx
@@ -1,0 +1,58 @@
+export const audioWave = (
+    <svg width="181" height="20" viewBox="0 0 181 20" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path d="M7.36914 15.793C6.9971 15.793 6.69678 15.4941 6.69678 15.1238V4.87623C6.69678 4.50594 6.9971 4.20703 7.36914 4.20703C7.74118 4.20703 8.04151 4.50594 8.04151 4.87623V15.1238C8.04151 15.4941 7.74118 15.793 7.36914 15.793Z" fill="#ACB3BA" />
+        <path d="M4.02051 14.1913C3.64847 14.1913 3.34814 13.8924 3.34814 13.5221V6.48218C3.34814 6.1119 3.64847 5.81299 4.02051 5.81299C4.39255 5.81299 4.69287 6.1119 4.69287 6.48218V13.5221C4.69287 13.8924 4.39255 14.1913 4.02051 14.1913Z" fill="#ACB3BA" />
+        <path d="M10.7173 14.1913C10.3452 14.1913 10.0449 13.8924 10.0449 13.5221V6.48218C10.0449 6.1119 10.3452 5.81299 10.7173 5.81299C11.0893 5.81299 11.3897 6.1119 11.3897 6.48218V13.5221C11.3897 13.8924 11.0893 14.1913 10.7173 14.1913Z" fill="#ACB3BA" />
+        <path d="M14.0659 12.7503C13.6939 12.7503 13.3936 12.4514 13.3936 12.0811V7.91871C13.3936 7.54842 13.6939 7.24951 14.0659 7.24951C14.438 7.24951 14.7383 7.54842 14.7383 7.91871V12.0811C14.7383 12.4514 14.438 12.7503 14.0659 12.7503Z" fill="#ACB3BA" />
+        <path d="M67.6309 15.793C67.2588 15.793 66.9585 15.4941 66.9585 15.1238V4.87623C66.9585 4.50594 67.2588 4.20703 67.6309 4.20703C68.0029 4.20703 68.3032 4.50594 68.3032 4.87623V15.1238C68.3032 15.4941 68.0029 15.793 67.6309 15.793Z" fill="#ACB3BA" />
+        <path d="M64.2827 14.1913C63.9107 14.1913 63.6104 13.8924 63.6104 13.5221V6.48218C63.6104 6.1119 63.9107 5.81299 64.2827 5.81299C64.6548 5.81299 64.9551 6.1119 64.9551 6.48218V13.5221C64.9551 13.8924 64.6548 14.1913 64.2827 14.1913Z" fill="#ACB3BA" />
+        <path d="M70.9795 14.1913C70.6075 14.1913 70.3071 13.8924 70.3071 13.5221V6.48218C70.3071 6.1119 70.6075 5.81299 70.9795 5.81299C71.3515 5.81299 71.6519 6.1119 71.6519 6.48218V13.5221C71.6519 13.8924 71.3515 14.1913 70.9795 14.1913Z" fill="#ACB3BA" />
+        <path d="M17.4141 14.0575C17.042 14.0575 16.7417 13.7586 16.7417 13.3883V6.61158C16.7417 6.24129 17.042 5.94238 17.4141 5.94238C17.7861 5.94238 18.0864 6.24129 18.0864 6.61158V13.3883C18.0864 13.7586 17.7861 14.0575 17.4141 14.0575Z" fill="#ACB3BA" />
+        <path d="M20.7627 15.5386C20.3907 15.5386 20.0903 15.2397 20.0903 14.8694V5.13038C20.0903 4.76009 20.3907 4.46118 20.7627 4.46118C21.1347 4.46118 21.4351 4.76009 21.4351 5.13038V14.8694C21.4351 15.2397 21.1347 15.5386 20.7627 15.5386Z" fill="#ACB3BA" />
+        <path d="M24.1064 17.3144C23.7344 17.3144 23.4341 17.0155 23.4341 16.6452V3.35499C23.4341 2.9847 23.7344 2.68579 24.1064 2.68579C24.4785 2.68579 24.7788 2.9847 24.7788 3.35499V16.6407C24.7788 17.011 24.4785 17.3099 24.1064 17.3099V17.3144Z" fill="#ACB3BA" />
+        <path d="M27.4551 15.6146C27.083 15.6146 26.7827 15.3157 26.7827 14.9454V5.05469C26.7827 4.68441 27.083 4.3855 27.4551 4.3855C27.8271 4.3855 28.1274 4.68441 28.1274 5.05469V14.9454C28.1274 15.3157 27.8271 15.6146 27.4551 15.6146Z" fill="#ACB3BA" />
+        <path d="M30.8032 14.0219C30.4312 14.0219 30.1309 13.723 30.1309 13.3527V6.65186C30.1309 6.28157 30.4312 5.98267 30.8032 5.98267C31.1753 5.98267 31.4756 6.28157 31.4756 6.65186V13.3527C31.4756 13.723 31.1753 14.0219 30.8032 14.0219Z" fill="#ACB3BA" />
+        <path d="M34.1519 13.1698C33.7798 13.1698 33.4795 12.8709 33.4795 12.5006V7.50391C33.4795 7.13362 33.7798 6.83472 34.1519 6.83472C34.5239 6.83472 34.8242 7.13362 34.8242 7.50391V12.5006C34.8242 12.8709 34.5239 13.1698 34.1519 13.1698Z" fill="#ACB3BA" />
+        <path d="M37.5 15.726C37.128 15.726 36.8276 15.4271 36.8276 15.0568V4.94752C36.8276 4.57723 37.128 4.27832 37.5 4.27832C37.872 4.27832 38.1724 4.57723 38.1724 4.94752V15.0568C38.1724 15.4271 37.872 15.726 37.5 15.726Z" fill="#ACB3BA" />
+        <path d="M50.8936 15.726C50.5215 15.726 50.2212 15.4271 50.2212 15.0568V4.94752C50.2212 4.57723 50.5215 4.27832 50.8936 4.27832C51.2656 4.27832 51.5659 4.57723 51.5659 4.94752V15.0568C51.5659 15.4271 51.2656 15.726 50.8936 15.726Z" fill="#ACB3BA" />
+        <path d="M40.8481 17.3144C40.4761 17.3144 40.1758 17.0155 40.1758 16.6452V3.35499C40.1758 2.9847 40.4761 2.68579 40.8481 2.68579C41.2202 2.68579 41.5205 2.9847 41.5205 3.35499V16.6407C41.5205 17.011 41.2202 17.3099 40.8481 17.3099V17.3144Z" fill="#ACB3BA" />
+        <path d="M44.1968 20C43.8247 20 43.5244 19.7011 43.5244 19.3308V0.669195C43.5244 0.298907 43.8247 0 44.1968 0C44.5688 0 44.8691 0.298907 44.8691 0.669195V19.3308C44.8691 19.7011 44.5688 20 44.1968 20Z" fill="#ACB3BA" />
+        <path d="M47.5449 17.3144C47.1729 17.3144 46.8726 17.0155 46.8726 16.6452V3.35499C46.8726 2.9847 47.1729 2.68579 47.5449 2.68579C47.917 2.68579 48.2173 2.9847 48.2173 3.35499V16.6407C48.2173 17.011 47.917 17.3099 47.5449 17.3099V17.3144Z" fill="#ACB3BA" />
+        <path d="M54.2417 13.1698C53.8697 13.1698 53.5693 12.8709 53.5693 12.5006V7.50391C53.5693 7.13362 53.8697 6.83472 54.2417 6.83472C54.6137 6.83472 54.9141 7.13362 54.9141 7.50391V12.5006C54.9141 12.8709 54.6137 13.1698 54.2417 13.1698Z" fill="#ACB3BA" />
+        <path d="M60.9385 13.1698C60.5664 13.1698 60.2661 12.8709 60.2661 12.5006V7.50391C60.2661 7.13362 60.5664 6.83472 60.9385 6.83472C61.3105 6.83472 61.6108 7.13362 61.6108 7.50391V12.5006C61.6108 12.8709 61.3105 13.1698 60.9385 13.1698Z" fill="#ACB3BA" />
+        <path d="M57.5903 12.3355C57.2183 12.3355 56.918 12.0366 56.918 11.6663V8.33814C56.918 7.96785 57.2183 7.66895 57.5903 7.66895C57.9624 7.66895 58.2627 7.96785 58.2627 8.33814V11.6663C58.2627 12.0366 57.9624 12.3355 57.5903 12.3355Z" fill="#ACB3BA" />
+        <path d="M0.672364 13.1698C0.300323 13.1698 0 12.8709 0 12.5006V7.50391C0 7.13362 0.300323 6.83472 0.672364 6.83472C1.04441 6.83472 1.34473 7.13362 1.34473 7.50391V12.5006C1.34473 12.8709 1.04441 13.1698 0.672364 13.1698Z" fill="#ACB3BA" />
+        <path d="M80.5657 15.793C80.1837 15.793 79.8754 15.4941 79.8754 15.1238V4.87623C79.8754 4.50594 80.1837 4.20703 80.5657 4.20703C80.9476 4.20703 81.256 4.50594 81.256 4.87623V15.1238C81.256 15.4941 80.9476 15.793 80.5657 15.793Z" fill="#ACB3BA" />
+        <path d="M77.1277 14.1913C76.7457 14.1913 76.4374 13.8924 76.4374 13.5221V6.48218C76.4374 6.1119 76.7457 5.81299 77.1277 5.81299C77.5096 5.81299 77.818 6.1119 77.818 6.48218V13.5221C77.818 13.8924 77.5096 14.1913 77.1277 14.1913Z" fill="#ACB3BA" />
+        <path d="M84.003 14.1913C83.6211 14.1913 83.3127 13.8924 83.3127 13.5221V6.48218C83.3127 6.1119 83.6211 5.81299 84.003 5.81299C84.385 5.81299 84.6933 6.1119 84.6933 6.48218V13.5221C84.6933 13.8924 84.385 14.1913 84.003 14.1913Z" fill="#ACB3BA" />
+        <path d="M87.441 12.7503C87.0591 12.7503 86.7507 12.4514 86.7507 12.0811V7.91871C86.7507 7.54842 87.0591 7.24951 87.441 7.24951C87.823 7.24951 88.1313 7.54842 88.1313 7.91871V12.0811C88.1313 12.4514 87.823 12.7503 87.441 12.7503Z" fill="#ACB3BA" />
+        <path d="M142.434 15.793C142.052 15.793 141.744 15.4941 141.744 15.1238V4.87623C141.744 4.50594 142.052 4.20703 142.434 4.20703C142.816 4.20703 143.125 4.50594 143.125 4.87623V15.1238C143.125 15.4941 142.816 15.793 142.434 15.793Z" fill="#ACB3BA" />
+        <path d="M138.997 14.1913C138.615 14.1913 138.307 13.8924 138.307 13.5221V6.48218C138.307 6.1119 138.615 5.81299 138.997 5.81299C139.379 5.81299 139.687 6.1119 139.687 6.48218V13.5221C139.687 13.8924 139.379 14.1913 138.997 14.1913Z" fill="#ACB3BA" />
+        <path d="M145.872 14.1913C145.49 14.1913 145.182 13.8924 145.182 13.5221V6.48218C145.182 6.1119 145.49 5.81299 145.872 5.81299C146.254 5.81299 146.563 6.1119 146.563 6.48218V13.5221C146.563 13.8924 146.254 14.1913 145.872 14.1913Z" fill="#ACB3BA" />
+        <path d="M149.31 12.7503C148.928 12.7503 148.619 12.4514 148.619 12.0811V7.91871C148.619 7.54842 148.928 7.24951 149.31 7.24951C149.692 7.24951 150 7.54842 150 7.91871V12.0811C150 12.4514 149.692 12.7503 149.31 12.7503Z" fill="#ACB3BA" />
+        <path d="M90.8784 14.0575C90.4964 14.0575 90.1881 13.7586 90.1881 13.3883V6.61158C90.1881 6.24129 90.4964 5.94238 90.8784 5.94238C91.2604 5.94238 91.5687 6.24129 91.5687 6.61158V13.3883C91.5687 13.7586 91.2604 14.0575 90.8784 14.0575Z" fill="#ACB3BA" />
+        <path d="M94.3164 15.5386C93.9344 15.5386 93.6261 15.2397 93.6261 14.8694V5.13038C93.6261 4.76009 93.9344 4.46118 94.3164 4.46118C94.6984 4.46118 95.0067 4.76009 95.0067 5.13038V14.8694C95.0067 15.2397 94.6984 15.5386 94.3164 15.5386Z" fill="#ACB3BA" />
+        <path d="M97.7493 17.3144C97.3673 17.3144 97.059 17.0155 97.059 16.6452V3.35499C97.059 2.9847 97.3673 2.68579 97.7493 2.68579C98.1312 2.68579 98.4395 2.9847 98.4395 3.35499V16.6407C98.4395 17.011 98.1312 17.3099 97.7493 17.3099V17.3144Z" fill="#ACB3BA" />
+        <path d="M101.187 15.6146C100.805 15.6146 100.497 15.3157 100.497 14.9454V5.05469C100.497 4.68441 100.805 4.3855 101.187 4.3855C101.569 4.3855 101.878 4.68441 101.878 5.05469V14.9454C101.878 15.3157 101.569 15.6146 101.187 15.6146Z" fill="#ACB3BA" />
+        <path d="M104.625 14.0219C104.243 14.0219 103.934 13.723 103.934 13.3527V6.65186C103.934 6.28157 104.243 5.98267 104.625 5.98267C105.007 5.98267 105.315 6.28157 105.315 6.65186V13.3527C105.315 13.723 105.007 14.0219 104.625 14.0219Z" fill="#ACB3BA" />
+        <path d="M108.063 13.1698C107.681 13.1698 107.372 12.8709 107.372 12.5006V7.50391C107.372 7.13362 107.681 6.83472 108.063 6.83472C108.445 6.83472 108.753 7.13362 108.753 7.50391V12.5006C108.753 12.8709 108.445 13.1698 108.063 13.1698Z" fill="#ACB3BA" />
+        <path d="M111.5 15.726C111.118 15.726 110.81 15.4271 110.81 15.0568V4.94752C110.81 4.57723 111.118 4.27832 111.5 4.27832C111.882 4.27832 112.19 4.57723 112.19 4.94752V15.0568C112.19 15.4271 111.882 15.726 111.5 15.726Z" fill="#ACB3BA" />
+        <path d="M125.251 15.726C124.869 15.726 124.56 15.4271 124.56 15.0568V4.94752C124.56 4.57723 124.869 4.27832 125.251 4.27832C125.633 4.27832 125.941 4.57723 125.941 4.94752V15.0568C125.941 15.4271 125.633 15.726 125.251 15.726Z" fill="#ACB3BA" />
+        <path d="M114.937 17.3144C114.556 17.3144 114.247 17.0155 114.247 16.6452V3.35499C114.247 2.9847 114.556 2.68579 114.937 2.68579C115.319 2.68579 115.628 2.9847 115.628 3.35499V16.6407C115.628 17.011 115.319 17.3099 114.937 17.3099V17.3144Z" fill="#ACB3BA" />
+        <path d="M118.375 20C117.993 20 117.685 19.7011 117.685 19.3308V0.669195C117.685 0.298907 117.993 0 118.375 0C118.757 0 119.066 0.298907 119.066 0.669195V19.3308C119.066 19.7011 118.757 20 118.375 20Z" fill="#ACB3BA" />
+        <path d="M121.813 17.3144C121.431 17.3144 121.122 17.0155 121.122 16.6452V3.35499C121.122 2.9847 121.431 2.68579 121.813 2.68579C122.195 2.68579 122.503 2.9847 122.503 3.35499V16.6407C122.503 17.011 122.195 17.3099 121.813 17.3099V17.3144Z" fill="#ACB3BA" />
+        <path d="M128.688 13.1698C128.306 13.1698 127.998 12.8709 127.998 12.5006V7.50391C127.998 7.13362 128.306 6.83472 128.688 6.83472C129.07 6.83472 129.378 7.13362 129.378 7.50391V12.5006C129.378 12.8709 129.07 13.1698 128.688 13.1698Z" fill="#ACB3BA" />
+        <path d="M135.563 13.1698C135.182 13.1698 134.873 12.8709 134.873 12.5006V7.50391C134.873 7.13362 135.182 6.83472 135.563 6.83472C135.945 6.83472 136.254 7.13362 136.254 7.50391V12.5006C136.254 12.8709 135.945 13.1698 135.563 13.1698Z" fill="#ACB3BA" />
+        <path d="M132.126 12.3355C131.744 12.3355 131.436 12.0366 131.436 11.6663V8.33814C131.436 7.96785 131.744 7.66895 132.126 7.66895C132.508 7.66895 132.816 7.96785 132.816 8.33814V11.6663C132.816 12.0366 132.508 12.3355 132.126 12.3355Z" fill="#ACB3BA" />
+        <path d="M73.6903 13.1698C73.3083 13.1698 73 12.8709 73 12.5006V7.50391C73 7.13362 73.3083 6.83472 73.6903 6.83472C74.0723 6.83472 74.3806 7.13362 74.3806 7.50391V12.5006C74.3806 12.8709 74.0723 13.1698 73.6903 13.1698Z" fill="#ACB3BA" />
+        <path d="M159.369 15.793C158.997 15.793 158.697 15.4941 158.697 15.1238V4.87623C158.697 4.50594 158.997 4.20703 159.369 4.20703C159.741 4.20703 160.042 4.50594 160.042 4.87623V15.1238C160.042 15.4941 159.741 15.793 159.369 15.793Z" fill="#ACB3BA" />
+        <path d="M156.021 14.1913C155.648 14.1913 155.348 13.8924 155.348 13.5221V6.48218C155.348 6.1119 155.648 5.81299 156.021 5.81299C156.393 5.81299 156.693 6.1119 156.693 6.48218V13.5221C156.693 13.8924 156.393 14.1913 156.021 14.1913Z" fill="#ACB3BA" />
+        <path d="M162.717 14.1913C162.345 14.1913 162.045 13.8924 162.045 13.5221V6.48218C162.045 6.1119 162.345 5.81299 162.717 5.81299C163.089 5.81299 163.39 6.1119 163.39 6.48218V13.5221C163.39 13.8924 163.089 14.1913 162.717 14.1913Z" fill="#ACB3BA" />
+        <path d="M166.066 12.7503C165.694 12.7503 165.394 12.4514 165.394 12.0811V7.91871C165.394 7.54842 165.694 7.24951 166.066 7.24951C166.438 7.24951 166.738 7.54842 166.738 7.91871V12.0811C166.738 12.4514 166.438 12.7503 166.066 12.7503Z" fill="#ACB3BA" />
+        <path d="M169.414 14.0575C169.042 14.0575 168.742 13.7586 168.742 13.3883V6.61158C168.742 6.24129 169.042 5.94238 169.414 5.94238C169.786 5.94238 170.086 6.24129 170.086 6.61158V13.3883C170.086 13.7586 169.786 14.0575 169.414 14.0575Z" fill="#ACB3BA" />
+        <path d="M172.763 15.5386C172.391 15.5386 172.09 15.2397 172.09 14.8694V5.13038C172.09 4.76009 172.391 4.46118 172.763 4.46118C173.135 4.46118 173.435 4.76009 173.435 5.13038V14.8694C173.435 15.2397 173.135 15.5386 172.763 15.5386Z" fill="#ACB3BA" />
+        <path d="M176.106 17.3144C175.734 17.3144 175.434 17.0155 175.434 16.6452V3.35499C175.434 2.9847 175.734 2.68579 176.106 2.68579C176.478 2.68579 176.779 2.9847 176.779 3.35499V16.6407C176.779 17.011 176.478 17.3099 176.106 17.3099V17.3144Z" fill="#ACB3BA" />
+        <path d="M179.455 15.6146C179.083 15.6146 178.783 15.3157 178.783 14.9454V5.05469C178.783 4.68441 179.083 4.3855 179.455 4.3855C179.827 4.3855 180.127 4.68441 180.127 5.05469V14.9454C180.127 15.3157 179.827 15.6146 179.455 15.6146Z" fill="#ACB3BA" />
+        <path d="M152.672 13.1698C152.3 13.1698 152 12.8709 152 12.5006V7.50391C152 7.13362 152.3 6.83472 152.672 6.83472C153.044 6.83472 153.345 7.13362 153.345 7.50391V12.5006C153.345 12.8709 153.044 13.1698 152.672 13.1698Z" fill="#ACB3BA" />
+    </svg>
+);
