+++ src/components/svg/playAudioIcon.tsx
@@ -1,0 +1,7 @@
+export const playAudioIcon = (
+    <svg width="16" height="17" viewBox="0 0 16 17" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path
+            d="M0.166504 6.05512V10.9449C0.166504 13.8783 0.166504 15.345 1.12472 15.9297C2.08294 16.5144 3.38726 15.8436 5.99588 14.502L5.9959 14.502L10.7498 12.0572C13.7789 10.4993 15.2935 9.72042 15.2935 8.5C15.2935 7.27958 13.7789 6.50067 10.7498 4.94285L5.9959 2.49797C3.38726 1.15639 2.08295 0.485599 1.12472 1.07031C0.166504 1.65503 0.166504 3.12173 0.166504 6.05512Z"
+            fill="#01A385" />
+    </svg>
+)
